# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: tempus API
  version: 0.1.0
  description: Coordinate + time feature queries with per-feature attribution.
servers:
  - url: /api/v1
paths:
  /query:
    get:
      summary: Query cached feature providers for a coordinate and time
      parameters:
        - {name: lat, in: query, required: true, schema: {type: number}}
        - {name: lon, in: query, required: true, schema: {type: number}}
        - {name: datetime, in: query, required: true, schema: {type: string}, description: "RFC3339 or YYYY-MM-DDTHH:MM[:SS]; no offset => UTC. Future instants are allowed: astronomy providers (sun/moon) compute for any date, while providers that cannot serve the future (weather) report a non-retryable error status in the envelope."}
        - {name: providers, in: query, required: false, schema: {type: string}, description: comma-separated provider ids}
        - {name: gddBase, in: query, required: false, schema: {type: number}, description: "Optional base temperature (°C, [-50,50]) for the aggregate provider's growing-degree-day accumulation. Omit to use the server default (10 °C)."}
        - {name: refPeriod, in: query, required: false, schema: {type: string}, description: "Optional reference period 'YYYY-YYYY' (start ≥ 1940) for the bioclim provider. Omit to auto-select the contemporaneous 30-year WMO normal (the period containing the datetime's year, or 1991-2020 for later years)."}
      responses:
        "200":
          description: Feature collection with per-provider status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryResult"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /query/batch:
    post:
      summary: Batch feature query for many coordinate+time points
      description: >
        Processes up to query.batch.max_points points through the same
        provider pipeline as GET /query. Responses stream as NDJSON (one
        result item per line, input order) when the client sends
        Accept: application/x-ndjson; otherwise a synchronous JSON envelope
        is returned, capped at query.batch.max_sync_points points (413 above
        that, with a hint to retry streaming). Invalid points become per-item
        error objects and never abort the batch. Batch traffic is
        rate-limited and budget-guarded against the Open-Meteo free tier;
        exhausted budget surfaces as retryable per-provider errors.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchQueryRequest"
      responses:
        "200":
          description: Batch results (sync envelope, or NDJSON stream of items)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchQueryResponse"
            application/x-ndjson:
              schema:
                $ref: "#/components/schemas/BatchQueryResultItem"
        "400":
          description: Invalid request (empty points, too many points, bad JSON)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: Body too large, or too many points for synchronous mode
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
    License:
      type: object
      required: [name, url, attribution]
      properties:
        name:
          type: string
        url:
          type: string
        attribution:
          type: string
    Feature:
      type: object
      properties:
        type:
          type: string
        geometry:
          type: object
          description: >
            GeoJSON geometry of the feature; always a Point for the current
            providers, carrying the coordinate the provider resolved the query
            to (which may differ from the requested one, e.g. a grid cell).
        properties:
          description: >
            The feature's properties. Which schema applies is decided by the
            kind property: the discriminator below maps each kind to the
            schema that documents it.
          oneOf:
            - $ref: "#/components/schemas/WeatherFeatureProperties"
            - $ref: "#/components/schemas/DewPointFeatureProperties"
            - $ref: "#/components/schemas/SunFeatureProperties"
            - $ref: "#/components/schemas/MoonFeatureProperties"
            - $ref: "#/components/schemas/AggregateFeatureProperties"
            - $ref: "#/components/schemas/BioclimFeatureProperties"
          discriminator:
            propertyName: kind
            mapping:
              weather: "#/components/schemas/WeatherFeatureProperties"
              dewpoint: "#/components/schemas/DewPointFeatureProperties"
              sun: "#/components/schemas/SunFeatureProperties"
              moon: "#/components/schemas/MoonFeatureProperties"
              aggregate: "#/components/schemas/AggregateFeatureProperties"
              bioclim: "#/components/schemas/BioclimFeatureProperties"
        license:
          $ref: "#/components/schemas/License"
    ProviderStatus:
      type: object
      properties:
        id:
          type: string
        kind:
          type: string
        status:
          type: string
          enum: [ok, unavailable, error]
        cached:
          type: boolean
        retryable:
          type: boolean
        retryAfter:
          type: string
          description: "Go duration string indicating how long to wait before retrying (e.g. \"30s\", \"2h0m0s\")"
        error:
          type: string
...
    BilingualLabel:
      type: object
      description: A label available in German and English.
      properties:
        de:
          type: string
          description: German label.
        en:
          type: string
          description: English label.
    FeaturePropertiesBase:
      type: object
      required: [kind]
      description: >
        Properties every feature carries, whatever its kind. The concrete
        schema is selected by kind (see Feature.properties).
      properties:
        kind:
          type: string
          description: >
            The feature's kind; the discriminator that selects the concrete
            property schema.
        provider:
          type: string
          description: >
            Identifier of the provider or deriver that produced the feature
            (e.g. "open-meteo", "dewpoint", "sun").
        observedAt:
          type: string
          format: date-time
          description: >
            The instant the feature describes (RFC3339, UTC). Absent on the
            bioclim feature, which describes a climate normal period and
            reports referencePeriod instead.
        units:
          type: object
          additionalProperties:
            type: string
          description: >
            Unit per numeric property of this feature, keyed by property name
            (e.g. {"temperature2m": "°C", "windSpeed10m": "km/h"}).
    WeatherFeatureProperties:
      description: >
        Properties of a weather feature (kind "weather") returned by the
        Open-Meteo provider. In addition to numeric measurements such as
        temperature2m, relativeHumidity2m, precipitation, windSpeed10m,
        cloudCover, and weatherCode, the feature carries isDay (computed from
        the sun's elevation angle via the NOAA approximation, not from the
        weather provider — Open-Meteo's archive is_day is unreliable for
        historical dates), isDaySource (citation for the solar computation), a
        bilingual human-readable explanation of the WMO weather code and its
        source citation, plus the Beaufort wind force derived from
        windSpeed10m.
      allOf:
        - $ref: "#/components/schemas/FeaturePropertiesBase"
        - type: object
          required: [kind]
          properties:
            kind:
              type: string
              enum: [weather]
            windBeaufort:
              type: integer
              minimum: 0
              maximum: 12
              description: >
                Beaufort wind force (0–12) derived from windSpeed10m using the
                published km/h class bounds (0: <1, 1: 1–5, 2: 6–11, 3: 12–19,
                4: 20–28, 5: 29–38, 6: 39–49, 7: 50–61, 8: 62–74, 9: 75–88,
                10: 89–102, 11: 103–117, 12: ≥118). Absent when windSpeed10m is
                missing, not finite, or reported without a unit or in an
                unrecognised one.
            windBeaufortDescription:
              # allOf, not a bare $ref: in OpenAPI 3.0.x a description sibling to
              # $ref is ignored by most tooling.
              allOf:
                - $ref: "#/components/schemas/BilingualLabel"
              description: >
                Bilingual (DE + EN) name of the Beaufort force (e.g.
                "frische Brise" / "Fresh breeze"); present whenever
                windBeaufort is present.
            windBeaufortSource:
              type: string
              description: >
                Source citation for windBeaufort, referencing "Beaufort-Skala
                (WMO), abgeleitet aus der Windgeschwindigkeit in 10 m, tempus".
            windBeaufortSourceURL:
              type: string
              format: uri
              description: >
                Canonical URL for the Beaufort scale, so a client can link the
                citation instead of only displaying it. Present whenever
                windBeaufortSource is.
            isDay:
              type: boolean
              description: >
                Whether it is daytime at the queried location and instant.
                Computed from the sun's elevation angle (NOAA Spencer approximation);
                true when the sun is above −0.833° (refraction + solar radius
                threshold). Not sourced from the weather provider.
            isDaySource:
              type: string
              description: >
                Citation for the isDay computation: "Tag/Nacht aus Sonnenstand
                berechnet (NOAA-Näherung), tempus".
            weatherCode:
              type: integer
              description: >
                WMO weather-interpretation code (Code Table 4677, Open-Meteo subset).
            weatherCodeDescription:
              allOf:
                - $ref: "#/components/schemas/BilingualLabel"
              description: >
                Bilingual (DE + EN) human-readable description of the WMO weather
                code; present when the code is a known WMO Code Table 4677 value.
            weatherCodeSource:
              type: string
              description: >
                Source citation for the weatherCodeDescription, referencing
                "WMO Code Table 4677 (WW – present weather); weather-interpretation
                codes as used by Open-Meteo".
            weatherCodeSourceURL:
              type: string
              format: uri
              description: >
                Canonical URL for the weather-code table, so a client can link
                the citation instead of only displaying it. Present whenever
                weatherCodeSource is.
    DewPointFeatureProperties:
      description: >
        Properties of a dew-point feature (kind "dewpoint") derived from the
        weather feature using the Magnus-Tetens formula. Includes the computed
        dew point, its unit, derivation metadata, and a bilingual comfort
        classification with source attribution.
      allOf:
        - $ref: "#/components/schemas/FeaturePropertiesBase"
        - type: object
          required: [kind]
          properties:
            kind:
              type: string
              enum: [dewpoint]
            dewPoint2m:
              type: number
              description: Dew point at 2 m height in °C (rounded to 0.1 °C).
            comfort:
              allOf:
                - $ref: "#/components/schemas/BilingualLabel"
              description: >
                Bilingual (DE + EN) human-comfort classification of the dew point
                (e.g. "sehr angenehm" / "very comfortable"). Based on the standard
                meteorological dew-point comfort scale.
            comfortSource:
              type: string
              description: >
                Source citation for the comfort classification, referencing the
                standard meteorological dew-point comfort scale.
            comfortSourceURL:
              type: string
              format: uri
              description: >
                Canonical URL for the dew-point comfort scale, so a client can
                link the citation instead of only displaying it. Present
                whenever comfortSource is.
    SunFeatureProperties:
      description: >
        Properties of a sun feature (kind "sun"), computed from the coordinate
        and instant via the NOAA solar approximation. Available for any date
        (past or future). All angles are in degrees; azimuth is measured
        clockwise from true north. Event times are RFC3339 UTC, or null when the
        event does not occur that day (polar day/night at the given threshold).
      allOf:
        - $ref: "#/components/schemas/FeaturePropertiesBase"
        - type: object
          required: [kind]
          properties:
            kind:
              type: string
              enum: [sun]
            elevationDeg:
              type: number
              description: Sun elevation above the horizon (degrees).
            azimuthDeg:
              type: number
              description: Sun azimuth, degrees clockwise from true north.
            zenithDeg:
              type: number
              description: Solar zenith angle (90 − elevation), degrees.
            lightPhase:
              allOf:
                - $ref: "#/components/schemas/BilingualLabel"
              description: >
                Bilingual light-regime label: day, civil / nautical / astronomical
                twilight, or night.
            sunrise:
              type: string
              nullable: true
              description: Sunrise (RFC3339 UTC) for the day, or null (polar day/night).
            sunset:
              type: string
              nullable: true
              description: Sunset (RFC3339 UTC) for the day, or null (polar day/night).
            solarNoon:
              type: string
              nullable: true
              description: Solar noon (RFC3339 UTC).
            solarNoonElevationDeg:
              type: number
              description: Sun elevation at solar noon (degrees).
            dayLengthMinutes:
              type: number
              description: Length of daylight (minutes) between sunrise and sunset.
            twilight:
              type: object
              description: >
                Civil (−6°), nautical (−12°) and astronomical (−18°) twilight, each
                with a dawn and dusk time (RFC3339 UTC or null).
              properties:
                civil: {type: object}
                nautical: {type: object}
                astronomical: {type: object}
    MoonFeatureProperties:
      description: >
        Properties of a moon feature (kind "moon"), computed from the coordinate
        and instant via Meeus's lunar theory (geocentric approximation).
        Available for any date. Angles in degrees, azimuth clockwise from true
        north. Rise/set times are RFC3339 UTC, or null when no crossing occurs.
      allOf:
        - $ref: "#/components/schemas/FeaturePropertiesBase"
        - type: object
          required: [kind]
          properties:
            kind:
              type: string
              enum: [moon]
            elevationDeg:
              type: number
              description: Moon elevation above the horizon (degrees, geocentric).
            azimuthDeg:
              type: number
              description: Moon azimuth, degrees clockwise from true north.
            distanceKm:
              type: number
              description: Earth–moon centre distance (km).
            illuminationPct:
              type: number
              description: Illuminated fraction of the moon's disc (percent).
            phaseAngleDeg:
              type: number
              description: Phase angle (degrees; 0 = full, 180 = new).
            ageDays:
              type: number
              description: Moon age since the last new moon (days).
            phase:
              allOf:
                - $ref: "#/components/schemas/BilingualLabel"
              description: >
                Bilingual phase name (Neumond … abnehmende Sichel / New Moon …
                Waning Crescent).
            moonrise:
              type: string
              nullable: true
              description: Moonrise (RFC3339 UTC), or null when the moon does not rise that day.
            moonset:
              type: string
              nullable: true
              description: Moonset (RFC3339 UTC), or null when the moon does not set that day.
    AggregateFeatureProperties:
      description: >
        Properties of a weather-aggregate feature (kind "aggregate"), computed
        by tempus from an Open-Meteo archive range: antecedent precipitation
        sums, the fund day's temperature extrema, growing-degree-days, and an
        Arrhenius thermal-time site index accumulated up to the instant. Needs
        past data — a future instant makes the provider report a non-retryable
        error.
      allOf:
        - $ref: "#/components/schemas/FeaturePropertiesBase"
        - type: object
          required: [kind]
          properties:
            kind:
              type: string
              enum: [aggregate]
            precipitation:
              type: object
              description: Precipitation sums (mm) ending at the instant.
              properties:
                last24h: {type: number, description: Sum over the preceding 24 h (mm).}
                last72h: {type: number, description: Sum over the preceding 72 h (mm).}
                last120h: {type: number, description: Sum over the preceding 5 days (mm).}
            temperatureDay:
              type: object
              description: The fund day's (UTC-day) temperature extrema in °C.
              properties:
                min: {type: number}
                max: {type: number}
                mean: {type: number}
            growingDegreeDays:
              type: object
              description: >
                Growing-degree-days accumulated from the start of the growing year to
                the instant (average method, floor 0, no upper cutoff). base5 and
                base10 are ALWAYS computed so a client cannot introduce inconsistency
                by silently changing the base over time; a per-request gddBase is
                added under custom.
              properties:
                base5: {type: number, description: "GDD at base 5 °C (°C·d)."}
                base10: {type: number, description: "GDD at base 10 °C (°C·d)."}
                custom:
                  type: object
                  description: GDD at the per-request gddBase (present only when gddBase is given).
                  properties:
                    baseCelsius: {type: number}
                    value: {type: number}
                since: {type: string, description: "Accumulation start date (YYYY-MM-DD): 1 Jan (northern) or 1 Jul (southern hemisphere)."}
                days: {type: integer, description: Number of days summed.}
                method: {type: string, description: GDD computation method.}
            arrheniusThermalTime:
              type: object
              description: >
                Species-agnostic, base-free site index: Boltzmann-Arrhenius thermal
                time accumulated to the instant, in reference-temperature-equivalent
                days (a day at referenceTempC contributes 1.0). Each day uses the
                2-point rate mean of Tmin and Tmax. Unlike a GDD base, referenceTempC
                is only a normalising constant, so values at different references are
                interconvertible.
              properties:
                value: {type: number, description: "Accumulated index (referenceTempC-equivalent days)."}
                referenceTempC: {type: number, description: "Normalisation reference temperature (°C), fixed at 20."}
                activationEnergyEv: {type: number, description: "Activation energy (eV), MTE value 0.65."}
                days: {type: integer, description: Number of days summed.}
                method: {type: string}
...
    QueryResult:
      type: object
      description: >
        Feature collection with per-provider status. The features[] array may
        include a weather feature (kind "weather") with WMO code descriptions
        (weatherCodeDescription DE/EN, weatherCodeSource) and a derived feature
        with kind "dewpoint" containing a dewPoint2m property (°C,
        Magnus-Tetens formula) plus a comfort classification (comfort DE/EN,
        comfortSource). It also includes a sun feature (kind "sun") and a moon
        feature (kind "moon"), computed for any date; see SunFeatureProperties
        and MoonFeatureProperties. It also includes a weather-aggregate feature
        (kind "aggregate"; see AggregateFeatureProperties) with antecedent
        precipitation, day temperature extrema and growing-degree-days, and a
        bioclim feature (kind "bioclim"; see BioclimFeatureProperties) with the
        19 WorldClim variables, the Köppen-Geiger class and the climate-derived
        altitudinal belt (Höhenstufe) + Rivas-Martínez thermotype. The providers[]
        array may include entries with ids "sun", "moon", "dewpoint",
        "aggregate" and "bioclim" reflecting their status; for future instants
        the "open-meteo" weather and "aggregate" providers report a
        non-retryable error while sun and moon remain "ok".
      properties:
        query:
          type: object
        features:
          type: array
          items:
            $ref: "#/components/schemas/Feature"
        providers:
          type: array
          items:
            $ref: "#/components/schemas/ProviderStatus"
    BatchQueryPoint:
      type: object
      required: [lat, lon, datetime]
      properties:
        id:
          type: string
          description: Opaque echo id; defaults to the 0-based input index as string.
        lat:
          type: number
        lon:
          type: number
        datetime:
          type: string
          description: Same formats as GET /query's datetime parameter.
    BatchQueryRequest:
      type: object
      required: [points]
      properties:
        providers:
          type: array
          items:
            type: string
          description: Optional provider-id filter applied to every point.
        gddBase:
          type: number
          description: Optional extra GDD base (°C, [-50,50]) applied to every point; base 5 and 10 are always included.
        refPeriod:
          type: string
          description: Optional bioclim reference period 'YYYY-YYYY' applied to every point.
        points:
          type: array
          items:
            $ref: "#/components/schemas/BatchQueryPoint"
    BatchQueryResultItem:
      description: >
        One result per input point, in input order: the single-query envelope
        plus id — or, for unparsable points, only id + error.
      allOf:
        - $ref: "#/components/schemas/QueryResult"
        - type: object
          required: [id]
          properties:
            id:
              type: string
            error:
              type: object
              properties:
                message:
                  type: string
    BatchQueryResponse:
      type: object
      required: [results, total, processing_time_ms]
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/BatchQueryResultItem"
        total:
          type: integer
        processing_time_ms:
          type: integer